make gcc 4.8.* happy

+++ src/ee/indexes/indexkey.h
@@ -158,7 +158,7 @@ public:
              * in this next uint64_t.
              */
             if (intraKeyOffset < 0) {
-                intraKeyOffset = sizeof(uint64_t) - 1;
+                intraKeyOffset = (int)sizeof(uint64_t) - 1;
                 keyOffset++;
             }
         }
@@ -174,7 +174,7 @@ public:
             retval |= (0xFF & (data[keyOffset] >> (intraKeyOffset * 8))) << (ii * 8);
             intraKeyOffset--;
             if (intraKeyOffset < 0) {
-                intraKeyOffset = sizeof(uint64_t) - 1;
+                intraKeyOffset = (int)sizeof(uint64_t) - 1;
                 keyOffset++;
             }
         }
@@ -184,7 +184,7 @@ public:
     std::string debug( const voltdb::TupleSchema *keySchema) const {
         std::ostringstream buffer;
         int keyOffset = 0;
-        int intraKeyOffset = sizeof(uint64_t) - 1;
+        int intraKeyOffset = (int)sizeof(uint64_t) - 1;
         const int columnCount = keySchema->columnCount();
         for (int ii = 0; ii < columnCount; ii++) {
             switch(keySchema->columnType(ii)) {
@@ -222,7 +222,7 @@ public:
         const TupleSchema *keySchema = tuple->getSchema();
         const int columnCount = keySchema->columnCount();
         int keyOffset = 0;
-        int intraKeyOffset = sizeof(uint64_t) - 1;
+        int intraKeyOffset = (int)sizeof(uint64_t) - 1;
         for (int ii = 0; ii < columnCount; ii++) {
             switch(keySchema->columnType(ii)) {
             case voltdb::VALUE_TYPE_BIGINT: {
@@ -260,7 +260,7 @@ public:
         ::memset(data, 0, keySize * sizeof(uint64_t));
         const int columnCount = keySchema->columnCount();
         int keyOffset = 0;
-        int intraKeyOffset = sizeof(uint64_t) - 1;
+        int intraKeyOffset = (int)sizeof(uint64_t) - 1;
         for (int ii = 0; ii < columnCount; ii++) {
             switch(keySchema->columnType(ii)) {
             case voltdb::VALUE_TYPE_BIGINT: {
